refactor(client/global utils): rename global utils to be *.util.ts instead of *.utils.ts
