+++ client/src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client'
 import { ApolloProvider } from '@apollo/client'
 
 import SettingsContextProvider from './contexts/SettingsContext.context.tsx'
-import { apolloClient } from './utils/apolloClient.utils.ts'
+import { apolloClient } from './utils/apolloClient.util.ts'
 import App from './App.tsx'
 
 import './index.css'
